refactor: extract virtual scrolling logic to external component and updated profile tabs accordingly
