--- shared/types/profile-tabs.ts
@@ -1,2 +1,0 @@
-export const validProfileTabs = ['', 'replies', 'likes'] as const;
-export type Tab = (typeof validProfileTabs)[number];
